functions/pyproject.toml
[project]
name = "functions"
version = "0.1.0"
description = "Lambda function handlers"
dependencies = [
    "sst",
    "fastapi==0.116.0",
    "mangum==0.19.0",
    "biopython==1.85",
    "python-jose[cryptography]==3.5.0",
    "pydantic==2.11.7",
    "pydantic-settings==2.10.1",
    "python-dotenv==1.1.1",
]
requires-python = "==3.12.*"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
sst = { git = "https://github.com/sst/sst.git", branch = "dev", subdirectory = "sdk/python" }
